requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
requests>=2.28.0
plotly>=5.0.0
gpt4all>=1.0.0
python-dotenv>=1.0.0 